read a frame transaction's frames from the transaction and its receipt
The el_tx_frames table held nothing that was not already in two other places: a
frame's target, value and gas budgets come from the transaction, and its result
from the receipt. Both outlive it - detailsRetention is clamped up to the
relational retention - so the table was a shorter-lived copy of longer-lived
data, and the only page that read it had already built the same frames from the
transaction before overwriting them with the rows.

The frames now come from the transaction and the receipt everywhere, which is
what the blockdb path already did. Where neither is retained the page says so
rather than showing a frame transaction with no frames.
