+++ handlers/debug_cache.go
@@ -221,7 +221,6 @@ var pruneObjectLabels = map[string]string{
 	txindexer.PruneObjTransactions:   "Transactions",
 	txindexer.PruneObjInternalTxs:    "Internal Txs",
 	txindexer.PruneObjTokenTransfers: "Token Transfers",
-	txindexer.PruneObjTxFrames:       "TX Frames",
 	txindexer.PruneObjBlocks:         "Blocks",
 	txindexer.PruneObjExecData:       "Exec Data",
 	txindexer.PruneObjTxHashIndex:    "TX-Hash Index",
