+++ indexer/execution/txindexer/pruning.go
@@ -14,7 +14,6 @@ const (
 	PruneObjTransactions   = "transactions"
 	PruneObjInternalTxs    = "internal_txs"
 	PruneObjTokenTransfers = "token_transfers"
-	PruneObjTxFrames       = "tx_frames"
 	PruneObjBlocks         = "blocks"
 	PruneObjExecData       = "exec_data"
 	PruneObjTxHashIndex    = "tx_hash_index"
